Unpublishing GitHub page and refreshing next package
Turns out that GitHub Pages can only host static pages. Which would be fine if my website was only using HTML, CSS, and JavaScript, or maybe a static framework for React, but the Next.js framework I am using with React runs a dynamic page. This does come with some backend opportunities that static pages don't have, but it would require me to set up a dedicated server for the website. At that point, it would be better to go through with my initial plan of hosting the website on AWS and ignore GitHub Pages.